♻️ refactor: styles 아래로 이동
